Refactor stripe trpc api
