+++ src/app/(main)/dashboard/billing/_components/manage-subscription-form.tsx
@@ -1,21 +1,18 @@
 "use client";
 
 import * as React from "react";
-import { type z } from "zod";
 
 import { Button } from "@/components/ui/button";
-import { type manageSubscriptionSchema } from "@/lib/validators/stripe";
+import type { ManageSubscriptionInput } from "@/server/api/routers/stripe/stripe.input";
 import { api } from "@/trpc/react";
 import { toast } from "sonner";
-
-type ManageSubscriptionFormProps = z.infer<typeof manageSubscriptionSchema>;
 
 export function ManageSubscriptionForm({
   isPro,
   stripeCustomerId,
   stripeSubscriptionId,
   stripePriceId,
-}: ManageSubscriptionFormProps) {
+}: ManageSubscriptionInput) {
   const [isPending, startTransition] = React.useTransition();
   const managePlanMutation = api.stripe.managePlan.useMutation();
